backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["suryaMadhav16 <rebbapragada.s@northeastern.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
beautifulsoup4 = "^4.12.3"
langchain = "^0.3.7"
langchain-openai = "^0.2.6"
openai = "^1.54.3"
playwright = "^1.48.0"
pypandoc = "^1.14"
pymupdf = "^1.24.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
